requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
python-dotenv
sqlalchemy          
mysql-connector-python
